+++ src/roq/gate_futures/protocol/json/parser.hpp
@@ -15,6 +15,8 @@
 #include "roq/gate_futures/protocol/json/tickers.hpp"
 #include "roq/gate_futures/protocol/json/trades.hpp"
 
+#include "roq/gate_futures/protocol/json/futures_system.hpp"
+
 namespace roq {
 namespace gate_futures {
 namespace protocol {
@@ -22,12 +24,14 @@ namespace json {
 
 struct Parser final {
   struct Handler {
-    virtual void operator()(Trace<protocol::json::Subscribe> const &) = 0;
-    virtual void operator()(Trace<protocol::json::Tickers> const &) = 0;
-    virtual void operator()(Trace<protocol::json::Trades> const &) = 0;
-    virtual void operator()(Trace<protocol::json::BookTicker> const &) = 0;
-    virtual void operator()(Trace<protocol::json::OrderBookUpdate> const &) = 0;
-    virtual void operator()(Trace<protocol::json::Candlesticks> const &) = 0;
+    virtual void operator()(Trace<Subscribe> const &) = 0;
+    virtual void operator()(Trace<Tickers> const &) = 0;
+    virtual void operator()(Trace<Trades> const &) = 0;
+    virtual void operator()(Trace<BookTicker> const &) = 0;
+    virtual void operator()(Trace<OrderBookUpdate> const &) = 0;
+    virtual void operator()(Trace<Candlesticks> const &) = 0;
+    //
+    virtual void operator()(Trace<FuturesSystem> const &) = 0;
   };
 
   static bool dispatch(Handler &, std::string_view const &message, core::json::BufferStack &, TraceInfo const &, bool allow_unknown_event_types);