+++ src/roq/gate_futures/protocol/json/trade_parser.cpp
@@ -92,6 +92,8 @@ bool TradeParser::dispatch(
       case ORDER_LIST:
         dispatch_helper<TradeOrderList>(handler, message, buffer_stack, trace_info);
         return true;
+      case FUTURES_SYSTEM:
+        break;
     }
   } else {  // subscribe or update do not have "header"
     switch (message_2.event) {
@@ -131,6 +133,7 @@ bool TradeParser::dispatch(
           case ORDER_CANCEL:
           case ORDER_CANCEL_CP:
           case ORDER_LIST:
+          case FUTURES_SYSTEM:
             break;
         }
         break;
@@ -167,6 +170,9 @@ bool TradeParser::dispatch(
           case ORDER_CANCEL_CP:
           case ORDER_LIST:
             break;
+          case FUTURES_SYSTEM:
+            dispatch_helper<FuturesSystem>(handler, message, buffer_stack, trace_info);
+            return true;
         }
         break;
       case API: