+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Changed
+
+* Support `futures.system` messages (#606)
+
 ## 1.1.5 &ndash; 2026-06-06
 
 ## 1.1.4 &ndash; 2026-04-20